// generated/src/app/api-models/stage-prem-pay-adjust-support.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StagePremPayAdjustSupport {
  batchId: string;
  transactionId: number;
  adjustId: number;
  arAdjCustomerId: string;
  arAdjStatementStatus: string;
  arAdjInvoiceNumber: number;
  arAdjPostDate: Date;
  arAdjCompanyCode: string;
  arAdjGlRefCode: string;
  arAdjTransactionDesc1: string;
  adjustmentReason: string;
  seqGroupId: number;
  groupType: string;
  groupId: string;
  subscriberId: string;
  debitGlNumber: string;
  creditGlNumber: string;
}